Final merge all nodes
